feat(trading-proto-upgrade): swap UTXO PoC using Storable State Machine (#1958)
This commit:
- Adds Storable State Machine abstraction with a goal to have as fewer changes to existing state machines as possible.
- Implements successful swap v2 of UTXO to UTXO coin. Adds tests for such swap using dockerized komodod daemons.
- Adds Swap V2 message exchange using Protobuf.

+++ mm2src/coins/lightning.rs
@@ -959,7 +959,7 @@ impl WatcherOps for LightningCoin {
     fn create_maker_payment_spend_preimage(
         &self,
         _maker_payment_tx: &[u8],
-        _time_lock: u32,
+        _time_lock: u64,
         _maker_pub: &[u8],
         _secret_hash: &[u8],
         _swap_unique_data: &[u8],
@@ -974,7 +974,7 @@ impl WatcherOps for LightningCoin {
     fn create_taker_payment_refund_preimage(
         &self,
         _taker_payment_tx: &[u8],
-        _time_lock: u32,
+        _time_lock: u64,
         _maker_pub: &[u8],
         _secret_hash: &[u8],
         _swap_contract_address: &Option<BytesJson>,
